Refactor SessionWindow sample
