+++ streaming/pom.xml
@@ -34,6 +34,7 @@
         <module>job-management</module>
         <module>kafka-source</module>
         <module>map-journal-source</module>
+        <module>session-windows</module>
         <module>socket-connector</module>
         <module>stock-exchange</module>
         <module>streaming-enrichment</module>
