+++ streaming/session-windows/src/main/java/com/hazelcast/jet/samples/sessionwindows/ProductEventType.java
