+++ streaming/session-windows/src/main/resources/log4j.properties
